Update cannot check dwp status page to match design
