+++ app/services/decisions/dwp_decision_tree.rb
@@ -6,18 +6,12 @@ module Decisions
         after_confirm_result
       when :confirm_details
         after_confirm_details
-      when :cannot_check_dwp_status
-        after_cannot_check_dwp_status
       else
         raise InvalidStep, "Invalid step '#{step_name}'"
       end
     end
 
     private
-
-    def after_cannot_check_dwp_status
-      edit('steps/client/has_benefit_evidence')
-    end
 
     def after_confirm_result
       if form_object.confirm_result.yes?
