+++ config/locales/en/steps.yml
@@ -88,6 +88,10 @@ en:
         edit:
           page_title: Client’s contact details
           heading: Enter your client’s contact details
+      cannot_check_dwp_status:
+        edit:
+          page_title: Sorry, we cannot check your client’s information with DWP
+          heading: Sorry, we cannot check your client’s information with DWP
 
     dwp:
       confirm_result:
@@ -101,10 +105,6 @@ en:
       benefit_check_result_exit:
         show:
           page_title: Use eForms for this application
-      cannot_check_dwp_status:
-        edit:
-          page_title: Sorry, we cannot check your client’s information with DWP
-          heading: Sorry, we cannot check your client’s information with DWP
 
     address:
       lookup:
